requirements/base.txt
Django==1.11.2
dj-database-url==0.4.2
django-webpack-loader==0.5.0
djangorestframework==3.6.3
gunicorn==19.7.1
psycopg2==2.7.1
python-decouple==3.0
requests==2.17.3
whitenoise==3.3.0